omx_venc: Handle output buffer parameter in secure mode
During secure encode usecase, though the output buffer
is allocated for the size equal to buffer metadata,
the alloc length, filled length and offset were updated
as per the bitstream data
Modify the encoder output buffer to be updated with
the metadata having the native handle. Secure encoder
clients has to extract the handle to get the required
information related to the buffer.

CRs-Fixed: 1041024
Change-Id: Ia6bd0010653d75751105892f36c52d5706a7d468